refactor(cli): flatten model command modules
Move the model and model-provider command slices to the neutral command root and update internal exports without changing public CLI or provider compatibility behavior.
